requirements.txt
paddlepaddle
fastdeploy-python
paddlenlp
datasets
numpy
loguru